ELEC-439: Made motor relay generic
